Removed Redis & Celery dependency from settings
The django-cache framework will use the default (memory)
cache for development since the CACHES setting was moved
to produciton.py. Also, the celery configuration along with
the celery rows in INSTALLED_APPS were moved to production.py.

Pls ignore wagtaildemo.db I moved it to the parent folder of
the repository.