Changed requirements to a directory
This follows a pattern introduced in two scoops of django in
order to have seperate requirements files for dev / prod. There
is a base.txt requirements files containing the reqs for both
environments and two more requirement files (dev.txt and
production.txt) containing the requirements for dev and prod.

The django-debug-toolbar has been moved to the dev while the
psycopg2, django-celery, gunicorn, redis and django-redis-cache
have been moved to the production.

Also the django-debug-toolbar has been moved to settings/dev.py
and the gunicorn INSTALLD_APP to the settings/prod.py